make story deletion link ask user for confirmation

+++ evently/backlog/unassigned/selectors/ol/_changes/selectors/a[href=#delete]/click.js
@@ -1,17 +1,24 @@
 function() {
   var a = $(this);
-  var id = a.parents('li').attr('data-id');
-  var db = $$(this).app.db;
-  db.openDoc(id, {
-    success: function(doc) {
-      $.log("Deleting doc:");
-      $.log(doc);
-      db.removeDoc(doc, {
-        success: function() {
-          $.log('Item deleted, yo.');
-        }
-      });
-    }
-  });
+  var parentLi = a.parents('li');
+  var storyName = $('.story_name', parentLi).text();
+  $.log('User clicked to delete story:');
+  $.log(storyName);
+  var message = "Delete story?\n" + storyName;
+  if (confirm(message)) {
+    var id = parentLi.attr('data-id');
+    var db = $$(this).app.db;
+    db.openDoc(id, {
+      success: function(doc) {
+        $.log("Deleting doc:");
+        $.log(doc);
+        db.removeDoc(doc, {
+          success: function() {
+            $.log('Item deleted, yo.');
+          }
+        });
+      }
+    });
+  }
   return false;
 }
